rename the main package
