+++ src/main/java/akademia/assistant/front/controller/Controller.java
@@ -1,0 +1,5 @@
+package akademia.assistant.front.controller;
+
+public interface Controller {
+    String getFXMLName();
+}
