+++ src/main/java/akademia/assistant/front/controller/MainWindowController.java
@@ -1,7 +1,7 @@
-package assistant.front.learningjavaassistantfrontend.controller;
+package akademia.assistant.front.controller;
 
-import assistant.front.learningjavaassistantfrontend.service.Service;
-import assistant.front.learningjavaassistantfrontend.view.ViewFactory;
+import akademia.assistant.front.service.Service;
+import akademia.assistant.front.view.ViewFactory;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
