+++ src/main/java/akademia/assistant/front/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package assistant.front.learningjavaassistantfrontend.exception;
+package akademia.assistant.front.exception;
 
 public class AuthenticationException extends RuntimeException {
     public AuthenticationException(String message) {
