+++ src/main/java/akademia/assistant/front/model/User.java
@@ -1,4 +1,4 @@
-package assistant.front.learningjavaassistantfrontend.model;
+package akademia.assistant.front.model;
 
 import java.io.Serializable;
