--- src/main/java/assistant/front/learningjavaassistantfrontend/controller/Controller.java
@@ -1,5 +1,0 @@
-package assistant.front.learningjavaassistantfrontend.controller;
-
-public interface Controller {
-    String getFXMLName();
-}
