+++ src/main/resources/akademia/assistant/front/view/main-window.fxml
