+++ src/main/resources/akademia/assistant/front/view/register-window.fxml
